advancedRagApp/requirements.txt
llama-index
weaviate-client
llama-index-vector-stores-weaviate
python-dotenv
deepseek-sdk
llama-index-llms-deepseek
llama-index-embeddings-huggingface
torch
sentence-transformers